# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Online Library API with CRUD operations.",
        "title": "Online Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/freebook": {
            "get": {
                "description": "Get list of all freebooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freebook"
                ],
                "summary": "Get all freebooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/freebook.FreeBook"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new freebook with file upload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freebook"
                ],
                "summary": "Create new freebook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Book Summary",
                        "name": "summary",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Author",
                        "name": "author",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Book Pages",
                        "name": "pages",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Cover Image",
                        "name": "cover_image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PDF File",
                        "name": "pdf_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/freebook/{id}": {
            "get": {
                "description": "Get a freebook by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freebook"
                ],
                "summary": "Get freebook by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/freebook.FreeBook"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update freebook details by ID (with optional new files)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freebook"
                ],
                "summary": "Update freebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Book Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Summary",
                        "name": "summary",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Author",
                        "name": "author",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Book Pages",
                        "name": "pages",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Cover Image",
                        "name": "cover_image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "PDF File",
                        "name": "pdf_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a freebook by ID",
                "tags": [
                    "freebook"
                ],
                "summary": "Delete freebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "freebook.FreeBook": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover_image": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "pdf_file": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}